简化 RCNNCore
