+++ interesting/testCCPD/RCNNCore/tools.py
@@ -1,0 +1,50 @@
+import cv2
+import torch
+import numpy as np
+
+# [京,沪,津,渝,鲁,冀,鄂,黑,苏,浙,皖,闽,赣,豫,粤,桂,琼,晋,蒙,辽,吉,云,藏,陕,甘,青,宁,湘,川,贵,新,港,澳,台]
+# [A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Q, R, S, T, U, V, W, X, Y, Z]
+# [0, 1, 2, 3, 4, 5, 6, 7, 8, 9]
+head = ["京", "沪", "津", "渝", "鲁", "冀", "鄂", "黑", "苏", "浙", "皖", "闽", "赣", "豫", "粤", "桂", "琼", "晋", "蒙", "辽", "吉", "云", "藏", "陕", "甘", "青", "宁", "湘", "川", "贵", "新", "港", "澳", "台"]
+letter = ["A", "B", "C", "D", "E", "F", "G", "H", "I", "J", "K", "L", "M", "N", "O", "P", "Q", "R", "S", "T", "U", "V", "W", "X", "Y", "Z"]
+numbers = ["0", "1", "2", "3", "4", "5", "6", "7", "8", "9"]
+
+words_list = head + letter + numbers + ["_", "$"]
+words_dict = { s:i for i,s in enumerate(words_list)}
+print("words_list: ", words_list)
+print("words_list len: ", len(words_list))
+
+def encode(str):
+    texts = []
+    for s in str:
+        if s in words_list:
+            texts.append(words_dict[s])
+        else:
+            texts.append(words_dict["_"])
+    return texts
+
+def decode(pred):
+    length = pred.size(0) #batch_size
+    
+    char_list = []
+    for i in range(length):
+        if pred[i] != len(words_list)-1 and pred[i] != len(words_list)-2 and (not (i > 0 and pred[i - 1] == pred[i])):
+            char_list.append(words_list[pred[i]])
+    print("char_list: ", char_list)
+    return ''.join(char_list)
+
+
+def load_image(imgpath):
+    img = cv2.imdecode(np.fromfile(imgpath, dtype=np.uint8), cv2.IMREAD_COLOR)
+    assert img is not None, 'Image Not Found ' + imgpath
+
+    # cv2.namedWindow("test", cv2.WINDOW_NORMAL)
+    # cv2.imshow("test", img)
+    # cv2.waitKey(0)
+
+    img = img[:, :, ::-1].transpose(2, 0, 1)
+    img = np.ascontiguousarray(img)
+
+    img = torch.from_numpy(img)
+    
+    return img
